revert the conversion of the bot face/vertex/normal arrays from being size_t back to int.  additionally, convert the remainder of bot struct size types over to size_t completely.  this propagates hundreds of ancillary changes (more than 400) but provides the added benefits of more extensive value range on some platforms, better warning/bug detection, and more consistently making size types be unsigned so no negative values is inherent.
svn:revision:42070
svn:branch:trunk
svn:account:brlcad

+++ include/wdb.h
@@ -202,14 +202,13 @@ WDB_EXPORT int mk_eto(
 WDB_EXPORT int mk_metaball(
     struct rt_wdb *wdbp,
     const char *name,
-    const int nctlpt,		/* number of control points */
+    const size_t nctlpt,	/* number of control points */
     const int method,		/* metaball rendering method */
     const fastf_t threshold,
     const fastf_t *verts[5] );	/* X, Y, Z, fldstr, goo/Beta */
 
-WDB_EXPORT WDB_EXTERN(int mk_arbn, (struct rt_wdb *fp, const char *name, int neqn, plane_t eqn[]) );
-WDB_EXPORT WDB_EXTERN(int mk_ars, (struct rt_wdb *fp, const char *name, int ncurves, int pts_per_curve,
-				   fastf_t	*curves[]) );
+WDB_EXPORT WDB_EXTERN(int mk_arbn, (struct rt_wdb *fp, const char *name, size_t neqn, plane_t eqn[]) );
+WDB_EXPORT WDB_EXTERN(int mk_ars, (struct rt_wdb *fp, const char *name, size_t ncurves, size_t pts_per_curve, fastf_t *curves[]) );
 
 /**
  * Given the appropriate parameters, makes the non-geometric
@@ -271,10 +270,10 @@ mk_bot(
     unsigned char	mode,
     unsigned char	orientation,
     unsigned char	error_mode,	/**<  may be used to indicate error handling (ignored for now) */
-    int		num_vertices,
-    int		num_faces,
+    size_t		num_vertices,
+    size_t		num_faces,
     fastf_t		*vertices,	/**<  array of floats for vertices [num_vertices*3] */
-    int		*faces,		/**<  array of ints for faces [num_faces*3] */
+    int			*faces,		/**<  array of ints for faces [num_faces*3] */
     fastf_t		*thickness,	/**<  array of plate mode thicknesses (corresponds to array of faces)
 					 * NULL for modes RT_BOT_SURFACE and RT_BOT_SOLID.
 					 */
@@ -288,19 +287,19 @@ mk_bot_w_normals(
     unsigned char	mode,
     unsigned char	orientation,
     unsigned char	flags,
-    int		num_vertices,
-    int		num_faces,
+    size_t		num_vertices,
+    size_t		num_faces,
     fastf_t		*vertices,	/**<  array of floats for vertices [num_vertices*3] */
-    int		*faces,		/**<  array of ints for faces [num_faces*3] */
+    int			*faces,		/**<  array of ints for faces [num_faces*3] */
     fastf_t		*thickness,	/**<  array of plate mode thicknesses (corresponds to array of faces)
 					 * NULL for modes RT_BOT_SURFACE and RT_BOT_SOLID.
 					 */
     struct bu_bitv	*face_mode,	/**<  a flag for each face indicating thickness is appended to hit point,
 					 * otherwise thickness is centered about hit point
 					 */
-    int		num_normals,	/**<  number of unit normals in normals array */
+    size_t		num_normals,	/**<  number of unit normals in normals array */
     fastf_t		*normals,	/**<  array of floats for normals [num_normals*3] */
-    int		*face_normals );	/**<  array of ints (indices into normals array), must have 3*num_faces entries */
+    int			*face_normals );/**<  array of ints (indices into normals array), must have 3*num_faces entries */
 
 /* brep.cpp */
 WDB_EXPORT int mk_brep( struct rt_wdb* wdbp, const char* name, ON_Brep* brep );
@@ -354,11 +353,11 @@ WDB_EXPORT void mk_pipe_init( struct bu_list *headp );
 
 /* strsol primitives */
 WDB_EXPORT WDB_EXTERN(int mk_dsp, (struct rt_wdb *fp, const char *name, const char *file,
-				   int xdim, int ydim, const matp_t mat));
+				   size_t xdim, size_t ydim, const matp_t mat));
 WDB_EXPORT WDB_EXTERN(int mk_ebm, (struct rt_wdb *fp, const char *name, const char *file,
-				   int xdim, int ydim, fastf_t tallness, const matp_t mat));
+				   size_t xdim, size_t ydim, fastf_t tallness, const matp_t mat));
 WDB_EXPORT WDB_EXTERN(int mk_vol, (struct rt_wdb *fp, const char *name, const char *file,
-				   int xdim, int ydim, int zdim, int lo, int hi,
+				   size_t xdim, size_t ydim, size_t zdim, size_t lo, size_t hi,
 				   const vect_t cellsize, const matp_t mat));
 WDB_EXPORT WDB_EXTERN(int mk_submodel, (struct rt_wdb *fp, const char *name, const char *file,
 					const char *treetop, int meth));
