+++ src/libwdb/ebm.c
@@ -39,7 +39,7 @@
  * M K _ E B M
  */
 int
-mk_ebm(struct rt_wdb *fp, const char *name, const char *file, int xdim, int ydim, fastf_t tallness, const matp_t mat)
+mk_ebm(struct rt_wdb *fp, const char *name, const char *file, size_t xdim, size_t ydim, fastf_t tallness, const matp_t mat)
     /* name of file containing bitmap */
     /* X dimansion of file (w cells) */
     /* Y dimension of file (n cells) */
